Refactors Node to be Decodable (#317)
* Refactors Node to be Decodable

This allows ContentPList to be decoded using a PropertyListDecoder. To
facilitate implementation of Decodable, I broke Node up into two parts:
Category and Sample. This more closely resembles the structure of the
plist.

* Decouples Sample from Category.

Also uses compiler-synthesized implementations of `Equatable` for `Category` and `Sample`.

* Replaces all remaining uses of `node` with `sample`.

* Moves decoding of categories into the app delegate

This is more of an app initialization thing than anything else. In addition, decoding the plist in `viewdidLoad()` could slow down presentation of the view.

* Adds documentation comments to samplesByNames(_:) function.

* Refactors decoding categories in app delegate

* Adds tests for decoding categories and samples.