Increase loading priority of bundle resource request
The documentation for `NSBundleResourceRequestLoadingPriorityUrgent` states the following: "A special value for loading priority informing the system that the user cannot continue until the resources marked with the tags managed by the request are downloaded." Since we don't create the resource request until the user has selected the sample, this seems like an appropriate priority for us to use. Hopefully this will result in users being able to get to the sample faster.